Implement RName() so that registers show local names.
Doing this requires re-creating the local stack of variables, and keeping it
up to date in accordance with the local variable scopes as we progress across
the various instructions.

One downside of this approach is that variable scopes almost never include the
instruction where they are initially assigned, which is usually the point
where you most want the name! But it's not that simple; oftentimes it'll be an
instruction like:

[05] add       2   2   258  ; R2(=foo) := R2 + K2(=1)

In this case the assignment belongs to the local variable, but other operands
also refer to R2, and they do *not* refer to the local, but an earlier
temporary value. So more complicated logic is required for correctly
displaying names at the point of assignment.